requirements.txt
django
djangorestframework
requests
djangorestframework-simplejwt